bootstrap: implement --snapshot-blob and --build-snapshot
This patch introduces --build-snapshot and --snapshot-blob options
for creating and using user land snapshots. At the time of the
creation of this patch, user land CJS modules and ESM are not
yet supported in the snapshot, but a subset of builtins should
already work (in particular modules loaded during bootstrap
are shipped in the built-in snapshot, so they should work
in user land snapshot as well), and support for more builtins
are being added.

To generate a snapshot using main.js as entry point and write
the snapshot blob to snapshot.blob:

$ node --snapshot-blob snapshot.blob --build-snapshot main.js

To restore application state from snapshot.blob:

$ node --snapshot-blob snapshot.blob